+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
-- Nothing here yet
+- Fixed HTML page title from "React App" to "NinjaTerm". 
 
 ## [4.0.0] - 2023-09-22