+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fixed HTML page title from "React App" to "NinjaTerm".
 - Updated logo in README to use standard red theme colour.
+- Added a snackbar to display status messages to the user.
+- Error handled correctly when user clicks the "Cancel" button in the serial port list.
 
 ## [4.0.0] - 2023-09-22