+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [4.1.0] - 2023-09-24
+
 ### Added
 
 - Added a snackbar to display status messages to the user.